+++ api/src/main/java/com/nexters/gaetteok/user/application/UserApplication.java
@@ -1,10 +1,10 @@
-package com.nexters.gaetteok.user.service;
+package com.nexters.gaetteok.user.application;
 
-import com.nexters.gaetteok.user.model.User;
+import com.nexters.gaetteok.domain.User;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserService {
+public class UserApplication {
 
     public User getUser(long id) {
         return null;
