+++ api/src/main/java/com/nexters/gaetteok/user/presentation/UserController.java
@@ -1,7 +1,7 @@
-package com.nexters.gaetteok.user.controller;
+package com.nexters.gaetteok.user.presentation;
 
-import com.nexters.gaetteok.user.controller.response.GetUserResponse;
-import com.nexters.gaetteok.user.service.UserService;
+import com.nexters.gaetteok.user.application.UserApplication;
+import com.nexters.gaetteok.user.presentation.response.GetUserResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
@@ -17,11 +17,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class UserController {
 
-    private final UserService userService;
+    private final UserApplication userApplication;
 
     @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<GetUserResponse> getUser(@PathVariable long id) {
-        return ResponseEntity.ok(GetUserResponse.of(userService.getUser(id)));
+        return ResponseEntity.ok(GetUserResponse.of(userApplication.getUser(id)));
     }
 
 }
