+++ settings.gradle
@@ -2,4 +2,7 @@ rootProject.name = 'gaetteok'
 
 include 'api'
 include 'infra'
+include 'domain'
+include 'domain:user'
+findProject(':domain:user')?.name = 'user'
